ngp ui and other code relating to ngp that isn't emulation

+++ src/burner/win32/burner_win32.h
@@ -301,7 +301,7 @@ bool MenuHandleKeyboard(MSG*);
 void MenuRemoveTheme();
 
 // sel.cpp
-extern int nLoadMenuShowX;
+extern UINT64 nLoadMenuShowX;
 extern int nLoadMenuShowY;
 extern int nLoadMenuExpand;
 extern int nLoadMenuBoardTypeFilter;
